fix(output): allow appending of file
Prior to this change if you used the same output filename as a previous
run it would happily overwrite it. This was less than ideal for
situations when the scanner crashed for some reason. This change makes
it so if you continue scanning you'll just append to the file and no new
header of column names will be written.

DCO-1.1-Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>